test: move getValue tests to express lib
